修改界面显示， 去除卡信息中部分显示。更改提示信息

+++ HandleRecordUI.xaml.cs
@@ -160,6 +160,7 @@ namespace ISSUDemoCrypto
 
         private void btn_Delete_Click(object sender, RoutedEventArgs e)
         {
+            if (dgrd_View.SelectedItem == null) return;
             dsHandleRecord record = (dsHandleRecord)dgrd_View.SelectedItem;
             if (HelperDatabase.DeleteRecord(record.MacId))
             {
